Time: 4 ms (42.00%) | Memory: 69.9 MB (10.75%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -1,7 +1,7 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double sum=0;
-        double max=Integer.MIN_VALUE;
+        long sum=0;
+        long max=Integer.MIN_VALUE;
         int i=0;
         for(int j=0;j<nums.length;j++)
         {
